+++ apps/tool_flask_generator/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ToolFlaskGeneratorConfig(AppConfig):
+    default_auto_field = 'flask.db.models.BigAutoField'
+    name = 'apps.tool_flask_generator'