+++ helpers/generator/__init__.py
@@ -14,6 +14,7 @@ from .generator_products  import *
 from .generator_user      import *
 from .generator_models    import *
 from .generator_api       import *
+from .dynamic_dt          import *
 from .generator_celery    import *
 from .generator_outh      import *
 
@@ -23,3 +24,5 @@ from .parser_deps         import *
 from .parser_env          import *
 from .parser_urls         import *
 from .parser_settings     import *
+
+from .generator_flask     import *